+++ app/models/research.rb
@@ -151,7 +151,10 @@ class Research < ApplicationRecord
       field :translations do
         label I18n.t('labels.translations')
       end
-      field :slideshows
+      field :slideshows do
+        partial "form_nested_many_sorting"
+        help I18n.t('admin.help.sort')
+      end
       field :related_items
     end
   end