+++ CHANGELOG.md
@@ -4,6 +4,7 @@ The numbers in brackets (i.e., [0.1]), indicate the version number of the app.
 ## [Unreleased]
 These are items that are have been finished, but not released yet.
 ### Added
+- Ability to control sort order by dragging/dropping items in News/Research Slideshows and Illustration Annotations
 - Change Log page to admin section so admins can see what is happening with the app
 
 ## [0.4] - 2019-01-21